requirements.txt
# MCP 核心依赖 - 使用FastMCP而不是mcp[cli]
fastmcp>=2.0.0

# 网络请求和浏览器自动化
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
selenium>=4.15.0
playwright>=1.40.0

# 数据处理
pydantic>=2.5.0
python-multipart>=0.0.6

# 认证和加密
cryptography>=41.0.0
pycryptodome>=3.19.0

# 文件和配置处理
python-dotenv>=1.0.0

# 日志
loguru>=0.7.2

# 开发和测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0 